requirements.txt
openai>=1.12.0
python-dotenv>=1.0.0
langchain>=0.1.12
langchain-openai>=0.0.8
duckduckgo-search>=4.4
browser-use>=0.1.40
PyPDF2>=3.0.0
pydantic>=2.0.0
langchain_google_genai
